Kafka/requirements.txt
flask 
pymongo 
flask-wtf 
wtforms
kafka-python